PKC: pkc host driver
C29XPCIE cards are a series of Freescale PCIe add-in cards working on
EP side to perform public key crypto acceleration. Here is the diagram
for PKC hardware setup and software components.

          RC (X86-64/PowerPC)            EP (C293PCIE card)
	-----------------------        -----------------------
	|   OpenSSL           |        |                     |
	|----------------------        |                     |
	|   CryptoDev         |        |   pkc-firmware      |
	|   CryptoAPI         |        |     (U-Boot)        |
	|   pkc-host driver   |        |                     |
	|                     |	 PCIe  |                     |
	|   PCIe data path    |<======>|   PCIe data path    |
	-----------------------        -----------------------

This patch is pkc-host driver running in the host side to offload the
encryption/decryption operations to EP side through PCIe data path.

pkc-firmware is the driver running in the EP side to push the
encryption/decryption operations from host to SEC engine and return
the results to host. It's maintained in another git tree.

Signed-off-by: Zhicheng Fan <B32736@freescale.com>
Signed-off-by: Mingkai Hu <Mingkai.Hu@freescale.com>