increase the number of available context pools for RSA operations
This patch targets a reduction in multi-core contention of spinlocks in
get_crypto_ctx and free_crypto_ctx. On a T1040QDS, 1400MHz host, this
results in an increase of performance from 256K to 340K RSA_PUB_1024 operations.

The same hack can be used for other operations as well. Currently,
there is no change in their behavior: the same context pool is used
as before (the one with index 0).

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>
Change-Id: I65ebda9fffaa8693a387100234b7bbfd794ea0e5
Reviewed-on: http://git.am.freescale.net:8181/37093