fix creation of circular list
If NUM_OF_RESP_RINGS is 1, the value of dev->fw_resp_rings[0].next is
never initialized in the loop and the second assignment turns into a
NOP since both lvalue and rvalue are the same.

When NUM_OF_RESP_RINGS is greater than one, the circular list ends up
leaving the first item [0] out of the chain by linking last item to
item [0].next which is item [1]

Fix both of these issues by chaining the last item with the first. If
NUM_OF_RESP_RINGS is one, we end up with a pointer to the same object.

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>
Change-Id: Ic60033f797b41b4e2a3d4f7fc139e9b56e421eae
Reviewed-on: http://git.am.freescale.net:8181/38564