remove "retry alloc" feature
This fragment addresses alloc failures only for a particular case
and does not consider recovery from failure or clean-up for the rest
of the code. It is never used or advertised as a feature and adds
burden for a clean-up fix that needs to be applied for this function.
Remove it for these reasons. Add it back properly if necessary.

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>
Change-Id: Ic53fd515e8060b118951829ca758a52082e3e460
Reviewed-on: http://git.am.freescale.net:8181/38702