partial revert of c_buffers references to fix digest regression
Digest and symmetric algorithms use a pointer indirection to locate the
actual buffers and since the buffer allocator did not account for this
indirection it resulted in a kernel oops when running digest operations.

This patch is a partial revert of 75d4ba9bd76895e7b59070a747c0d41e2fb38d3a
and adds back the pointer indirection to the operation buffers to fix the error.

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>
Change-Id: Id0877e118e0df3c333a28e08fa7acc07190019d2
Reviewed-on: http://git.am.freescale.net:8181/38832