fix isr_context memory deallocation
The isr_context memory allocated in bulk before the loop was
incorrectly freed at driver removal. This patch fixes also the
deallocation on error path for the entire loop and makes a series of
code modifications to help with that. In particular, isr_context is
added to the list only if everything in the current iteration succeeds.

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>
Change-Id: I9b0dbd53026794446210f752446ed9910047da91
Reviewed-on: http://git.am.freescale.net:8181/39455