config device for proper module removal
In addition to resetting the core the CPU0_EN field of the
ECM_EEBPCR register in the devices CCSR must be reset. This
will block the core's access to the boot vector untill the same
field is set again (which is done in the init module function).

In the actual configuration, since the field is set and never
reset, once the device core is enabled (during the next second
insmod), it starts booting the old firmware accesing the previous
OB (out-of-band) memory for the handshake. The HOST allocates a
new OB memory buffer every time the driver is loaded but fais to
sends the new base address for the OB memory in time. Therefore
it is created a situation where the device writes on a different
memory page then the host expects to read.

Signed-off-by: Radu Alexe <raduandrei.alexe@freescale.com>