always link pkc.o when building the driver
The local pkc.c sources provide symbols not yet available in upstream Linux
kernels but present in Freescale SDK Linux. These symbols are necessary
for pkc support which is not build-time optional.

Since the symbols are always necessary, this makes the EXTRA_PKC flag
redundant since it must always be set. We remove it to uniformize
driver building on all platforms.

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>