fix calculation of outbound memory length
The necessary memory is currently below two cachelines and there is
no stability issue. However, this assumption is incorrect and the code
needs to account for the actual structure size.
This change does not affect the next field which is still cacheline
alligned to avoid cache thrashing.

The two structure members are not actually used and we can safely remove
them.

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>