remove unnecessary lock contention
The context for the current operation is used during irq handling
to finish the operation and perform any clean-up on associated data and
itself. This private data/crypto context was originally obtained from
crypto context pool and will be released after irq handling back to the
pool. Only these operations require locks on crypto context pool.

Saving and retrieving a pointer to this crypto context is just like any
other operation on crypto context and does _not_ require a lock,
especially not one on context pool!

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>