clear ring lists before freeing ring memory
When the crypto layer is initialised, the memory for the rings is
allocated and then, for each core, the ring lists are created. At
driver exit the operations must be reversed: the ring lists have
to be cleared and then the memory is freed.

This fix makes unnecessary deleting list heads in "cleanup_percore_list()".

Signed-off-by: Radu Alexe <raduandrei.alexe@freescale.com>