refactoring: use sg_cnt as index
Although variable i has the same value as sg_cnt when for loop
terminates, this confuses KW which reports a possible out of bound
access on src_sg array.
This function is called when sg_cnt is not NULL so we do not risk
the out of bound error anyway.

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>