execute device reset inside device remove function
There is no reason to bulk reset all devices in driver exit function. In
fact, if /sys/bus/pci filesystem is used to control the device from
user-space, the device may be removed from the system while still
running. This is of no importance in itself but insmod will fail because
the device is not in the expected hold-off state.

By itself this patch does not fix the above probing issue because the
CPU core is not put in hold-off mode with the current procedure. It is
however a necessary step for the next clean-up patches.

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>
Reviewed-by: Radu Alexe <raduandrei.alexe@freescale.com>