fix CPU core reset sequence
This patch fixes several handshake issues related to driver reloading.

At power-on, the CPU core is in hold-off mode and this allows firmware
to be uploaded. If the driver is reinserted, the CPU must be stopped
before new firmware is loaded again.

It is not possible to put the CPU back into hold-off mode unless a hard
reset is asserted. If issued, a hard reset will trigger also PCIe bus
errors on the host and will affect all device subsystems - including
PCIe registers which will need to be restored. Hard reset is also the
only way to bring back some subsystems if they are put in a low power
mode.

In this patch we issue just a CPU core reset instead of a hard reset since
the power management features are not used and to avoid losing PCIe
configuration data.

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>
Signed-off-by: Radu Alexe <raduandrei.alexe@freescale.com>