decrement device count on device remove
If we keep removing and reinserting a device by /sys/bus/pci means,
the device count will keep growing and we end up with runtime memory
leaks because new configurations are being added.

This patch fixes the issue when one device is present on the system but
the current design is not appropriate for multiple devices.

This decrement was lost in c31be2992908ce00a0bfb93e6421eba0cb58ecfe

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>