split boot procedure into independent functions
This patch limits responsibilities of several functions to specific
tasks. The device is now explicitely stopped before we change it's
configuration registers as required by LAW update procedure.

The assumption that the core is stopped at the configuration point
is wrong for the case when driver's probe function is interrupted
and then restarted.

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>