remove redundant structure ctx
This wrapper is used just as an enclosure over crypto_op_ctx and adds no
benefit to the code. The utility functions are simplified to refer
directly the crypto_op_ctx structures as they are of interest to the
callers.

To remove this wrapper we just move the 'next' pointer from ctx_t to
crypto_op_ctx and the rest of modifications follow.

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>