fix reading values from the local memory
This is host memory where the device has already written values in
big endian. The device is no longer concerned with these values and
any changes by the host are not necessary - especially if they are
in host endiannes.

On the other hand "fixing" the value for the host by rewriting it in
host endianness for the next reader poses the problem of making sure the
"fix" was already applied. This makes the code fragile and adds
inconsistencies. Instead, we should simply read these values and convert
them on the fly whenever necessary.

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>