make volatile the handshake memory on host
This is where the device writes to and we must avoid silent compiler
optimizations. Since it is local memory we simply read it and convert
the values written by the device from big to host endian. We do not
save the converted values to avoid relying on the conversion when a
different reader wants the same value.

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>