use dev_err to print device related errors
The purpose of this patch is to remove error prints dependency on
compile time options. While debug prints are not particular important
on live systems, leaving out error prints leads to untraceable issues
when they appear in the field. Not having any error logging is not
acceptable. This patch is first in a series that transitions to linux
methods for debug and error printing.

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>