drop volatile qualifier for device handshake area
The volatile qualifier was overridden anyway by casts to (void *). It's
purpose however was to ensure that all necessary data is present in the
expected structures when the host issues commands to the device.
This is better served by ensuring data ordering. Thus, we issue a memory
barrier just before changing the state of the device, to make sure that
data writes are flushed before it.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>