fix host_to_dev conversion
This patch breaks backward code compatibility. A series of supporting
patches are required as well.

The essence of this patch is that we will communicate to the firmware
host data pointers that are based on DMA host addresses instead of
physical ones. This change is necessary to be able to run the driver
onto kernels where IOMMU is enabled.

- descriptors are located in host ip_pool which is already DMA mapped.
  It's offset into the ip_pool is added to the ip_pool dma address to
  obtain the dma address of the descriptor
- input and output buffers are located anywhere in host RAM and we need
  to DMA map them before we can use them

The host to device conversion simply applies an offset to the host dma
address to obtain the C293 physical address as required by the SEC engine
(36 bits). The C293 virtual address as required by the CPU has different
mapping and is not useful except for debug purposes.

For reference these are the address conversion functions:

h_dma_addr = host_pci_map(h_v_addr)
d_p_addr = h_dma_addr            + 0x800000000  ; used by C293 SEC
d_v_addr = d_p_addr & 0x3fffffff +  0x80000000  ; used by C293 CPU

With this patch we no longer allow descriptors into the device SRAM (by
memcopy means) because it complicates host2dev function. Besides, the
whole concept of seeing device memory as a shadow of the host one lead
to a bunch of code review confusions. If we need to put data in device
SRAM we should make that explicit

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>