only high perf mode is supported
high-perf mode is nothing more than the normal execution mode for the
driver - where no "commands" can be sent to the firmware. Since the
non-high-perf mode requires firmware built with the same flags and
because this "command" mode has never been tested or supported, we
remove the option from Makefile to avoid confusion. The conditional code
will be removed as well in next patches.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>