remove custom descriptor address conversion
The conditional was based on the fact that descriptor address was
located in device SRAM (0xfff00000-0xffffffff) or in device's PCI window
(0x800000000 and above). Since we no longer run descriptors from device
memory, we can remove this check and use a single conversion function to
find out the descriptor's address in host memory.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>