remove unnecessary check on pointer before free
With the previous fixes on error path clean-up code, dealloc_crypto_mem
is called only when alloc_crypto_mem succeeded. Thus, it receives valid
pointers which don't need to be checked before use.

The buffer free is performed last for clarity, in reversed order in which
allocation and buffer mapping is being performed.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>