disable some debug print messages
Some messages classified as errors were enabled with
143d79f3063c49689d1d64d852cb64f06f14ad79 do not silence print_error at build time

However, not all error messages are errors and these ones are only info
or debug messages at most. We silence them to avoid flooding the console
on high loads (see DH_TEST_1K) and we add a note that the driver needs
improvement with a back-off mechanism.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>