fix incorrect DMA access to kernel memory in ecdh tests
Several tests suffer from this problem. With the device accessing host
memory, that memory should be allocated with kmalloc(GFP_DMA) to avoid
access errors. For some reason, on ARM and powerpc, memory located in
kernel space (global module variables) fail to be visible by the device.
We fix this problem by allocating all the memory that the device need to
access during tests.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>