remove specialized macro dev_print_dbg
While associating the debug message with a particular device driver is
useful, currently this will not give any information because the
driver supports only one device.
We remove this debug print macro to simplify the code and allow for a
more clean approach in the future if this proves necessary.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>