remove not used asignment
How come the length of the string was the sizeof int?! Remove that.

We're writing a 32 bit number so we have at most 10 chars in that
buffer. Now, do we have at least 11 bytes in that buffer? Never mind the
buffer string copy in the other branch. That is a question for you to
answer :)

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>