merge two related structures into one
The reason for this merge is to simplify code structure. There is no
architectural reason for the split between these two structures and
no useful abstraction that could convice to keep them separate. Since
both structures had pointers to each other, the transformation is
trivial and the result is refreshing.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>