fix issues with install target
Aparently, on some setups, 'sudo make' will clear the $PWD variable
on the first Makefile expansion. This leads to failures of "sudo make
install" or other issues when trying to preserve the environment with
sudo -E.

There are two solutions to this problem:
1) Use $$PWD instead of $(PWD) to render `$PWD` as the actual string to
be used on the secondary Makefile expansion.
This is used (albeit inconsistently) in the Linux kernel guide for
external modules:
kernel https://www.kernel.org/doc/Documentation/kbuild/modules.txt

2) A second solution is based on GNU make which sets CURDIR to the
pathname of current working directory. This variable is never touched
again by make.
This solution is choosen for it is just as clear as the PWD one would
have been had it worked in the first place.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>