make ob_mem a local variable
Ob_mem is a temporary storage and there is no need to keep it longer
than needed. Only two offsets are derivated later and the long term
solution is to clean-up the code. For now we maintain backward
compatibility with rings and ip_pool by recalculating the offset of
the data that was kept in ob_mem.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>