avoid indirect access to buf_pools through rings
Currently a single buf_pool is shared by all rings and this is a problem
because of spinlocks. Since in the end we want to have per-core buffer
pools we need first to remove association with the rings.

In a first phase we'll have an upper limit of buffer pools equal with the
number of ring pairs but this limit will be lifted later.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>